backend/requirements.txt
asgiref==3.5.0
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
mpmath==1.2.1
numpy==1.21.5
pytz==2022.1
sqlparse==0.4.2
sympy==1.10.1
typing-extensions==4.1.1
